requirements.txt
Django==1.10.5
psycopg2==2.7
djangorestframework==3.6.2